work_shop_agent_ai/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
